+++ metacatalog/ext/standard_export/extension.py
@@ -1,14 +1,15 @@
 from typing import Union
 
 
-from lxml import etree
-import xmltodict
+import os
+from tqdm import tqdm
+from sqlalchemy.orm import Session
 
 
 from metacatalog.ext import MetacatalogExtensionInterface
-from metacatalog.ext.standard_export.util import _parse_iso_information, _init_iso19115_jinja, _validate_xml
+from metacatalog.ext.standard_export.util import _parse_iso_information, _init_iso19115_jinja, _validate_xml, _get_uuid
 from metacatalog import api
-from metacatalog.models import Entry, Person
+from metacatalog.models import Entry
 from metacatalog.util.results import ImmutableResultSet
 
 
@@ -20,14 +21,25 @@ class StandardsExportExtension(MetacatalogExtensionInterface):
     @classmethod
     def init_extension(cls):
         # wrapper which calls StandardsExportExtension.iso19115_export
-        def wrapper(self: Entry, config_dict: dict, path: str = None):  
+        def wrapper_entry(self: Entry, config_dict: dict, path: str = None):  
             return StandardsExportExtension.iso19115_export(entry_or_resultset=self, config_dict=config_dict, path=path)
         
         # iso19115_export docstring and name for wrapper function
-        wrapper.__doc__ = StandardsExportExtension.iso19115_export.__doc__
-        wrapper.__name__ = StandardsExportExtension.iso19115_export.__name__
+        wrapper_entry.__doc__ = StandardsExportExtension.iso19115_export.__doc__
+        wrapper_entry.__name__ = StandardsExportExtension.iso19115_export.__name__
         
-        Entry.export_iso19115 = wrapper
+        # add wrapper to Entry model
+        Entry.export_iso19115 = wrapper_entry
+
+        # add function create_iso19115 to api.catalog
+        def wrapper_api(session: Session, config_dict: dict, path: str, if_exists: str = 'fail', verbose: bool = False):
+            StandardsExportExtension.create_iso19115(session, config_dict, path, if_exists, verbose)
+
+        wrapper_api.__doc__ = StandardsExportExtension.create_iso19115.__doc__
+        wrapper_api.__name__ = StandardsExportExtension.create_iso19115.__name__
+
+        # add wrapper to api.catalog
+        api.catalog.create_iso19115 = wrapper_api
 
 
     @classmethod
@@ -82,3 +94,68 @@ class StandardsExportExtension(MetacatalogExtensionInterface):
         else:
             with open(path, 'w') as f:
                 f.write(xml)
+
+
+    def create_iso19115(session: Session, config_dict: dict, path: str, if_exists: str = 'fail', verbose: bool = False) -> None:
+        """
+        Generate ISO 19115 XML files for all ImmutableResultSets in the
+        database session. The XML files are saved in the folder given in
+        ``path``, existing files in the folder are deleted, so use this 
+        function with caution.
+
+        .. versionadded:: 0.7.4
+
+        Parameters
+        ----------
+        session : sqlalchemy.Session
+            SQLAlchemy session connected to the database.
+        config_dict : dict
+            Configuration dictionary, containing information about the data provider
+        path : str
+            Folder location where all ISO19115 XML files are saved to.
+        if_exists: {'fail', 'replace'}, default 'fail'
+            How to behave if the XML file for the ImmutableResultSet already exists in path.
+
+            * fail: Raise a ValueError
+            * replace: Overwrite the existing XML file.
+        verbose: bool, default False
+            Enable verbose output.        
+
+        """
+        from metacatalog.ext.standard_export.util import _get_uuid
+        from tqdm import tqdm
+
+        if if_exists not in ("fail", "replace"):
+            raise ValueError(f"'{if_exists}' is not valid for if_exists")
+
+        irs_uuids = []
+
+        # use absolute path
+        path = os.path.abspath(path)
+
+        # list files to check if a file already exists
+        files = os.listdir(path)
+
+        # create the generator 
+        if verbose:
+            gen = tqdm(api.find_entry(session))
+        else:
+            gen = api.find_entry(session)
+
+        for entry in gen:
+            # get the uuid of the ImmutableResultSet that is written to ISO19115 XML (rs.group.uuid or rs.get('uuid'))
+            irs_uuid = _get_uuid(ImmutableResultSet(entry))
+
+            # if irs_uuid in irs_uuids: ImmutableResultSet already exported -> continue
+            if irs_uuid in irs_uuids:
+                continue
+            else:
+                # check if_exists policy first
+                if any(irs_uuid in file for file in files):
+                    if if_exists == 'fail':
+                        raise ValueError(f"ISO19115 XML file for uuid '{irs_uuid}' already exists under {path}.")
+                
+                entry.export_iso19115(config_dict, path=f"{path}/iso19115_{irs_uuid}.xml")
+
+                irs_uuids.append(irs_uuid)
+
